requirements.txt
BeautifulSoup
requests
fake-useragent
python-docx